requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comet-ml==3.33.3
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dulwich==0.21.5
entrypoints==0.4
everett==3.1.0
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyarrow==12.0.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.3.5
semantic-version==2.10.0
sentry-sdk==1.24.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.20.0
websocket-client==1.3.3
Werkzeug==2.3.4
wrapt==1.15.0
wurlitzer==3.0.3
zipp==3.15.0
